Working usefulness functionality
Was actually working before, just wasn't apparent as the bias of the
final convolutional layer was 0 and thus in printing showed 0's. Fixed
up the logic to be in function definitions instead of giant
lambdas. Also, was able to remove the num_classes and num_features
parameters and infer them from the data and model.