Feature/feature classifier (#41)
* WIP. Added the Feature Classifier Mixin

WIP commit. Implemented the initial FeatureClassifierMixin class with
usefulness implemented. Current bugs exist in the usefulness method,
including oddities with an all zeros correltation tensor (probably
something to do with int vs float or with the class_index method).

* Working usefulness functionality

Was actually working before, just wasn't apparent as the bias of the
final convolutional layer was 0 and thus in printing showed 0's. Fixed
up the logic to be in function definitions instead of giant
lambdas. Also, was able to remove the num_classes and num_features
parameters and infer them from the data and model.

* Add robustness calculation to FeatureClassifierMixin

* Fix travis with formatting & dm-sonnet dependency

* Get FeatureClasifier.robust working

- Add a test directory to .gitignore
- Remove accidentally hardcoded values in FeatureClassifierMixin.robustness
- Added ReLUs to our FeatureClassifierTest model to ensure that the fast gradient sign method actually works consistently
- Added model caching to the FeatureClassifierTest for consistent results
- Add seeds to shuffle in setup_tfds to give the option to remove that factor of variation

* Fixed a dumb

* Giving up on getting travis to be deterministic

* Fill some docs, small adjustments to adversarial

* Update CODEOWNERS

* Fix a small doc wording issue

Co-authored-by: Eleanor Quint <DrKwint@gmail.com>